fix(Jenkinsfile): update Docker image name
The Jenkinsfile has been modified to update the Docker image name used for building and pushing the image. The previous image name "josenerydev/WeatherForecast" has been replaced with "josenerydev/weather-forecast" for better clarity and consistency. This change ensures that the naming convention for the Docker image is aligned with the updated requirements. This commit addresses the core components affected by these alterations.

BREAKING CHANGE: None