feat(Jenkinsfile): streamline Docker image build and push
Simplify the Jenkinsfile by removing comments and unused code related to building and testing a Docker image. Focus on pushing the Docker image to Docker Hub. This change aims to streamline the pipeline, making it more efficient. By removing the build and test stages, the process becomes faster and more straightforward. This can result in quicker feedback and deployment of the Docker image.