feat(Jenkinsfile): update Docker build command
The Docker build command in the 'Build Docker Image' stage of the Jenkinsfile has been updated to target the directory "src/WeatherForecast.WebApi/" instead of "src/WeatherForecast/". This change indicates that the modifications made in the code impact the WeatherForecast component, specifically its WebApi. The purpose of this change is to build and deploy the Docker image for the updated WeatherForecast.WebApi component.